# Sg/do time step as autograd function