--- a/emu_base/math/double_krylov.py
+++ b/emu_base/math/double_krylov.py
@@ -88,7 +88,6 @@ def lanczos(
 
         if n2 < tolerance:
             converged = True
-            print("happy")
             break
 
         lanczos_vectors.append(w / n2)