… test/emu_base/math/test_double_krylov.py
… test/emu_sv/test_dense_operator.py