Support gradient backpropagation for time steps where $\Omega$, $\Delta$ or $\Phi$ require it.

### Changes
- `do_time_step` is now a `torch.autograd.Function`
    - forward simply do a krylov time step as before
    - backward propagate the gradient of the output state to the input (as prescribed in [torch docs](https://pytorch.org/tutorials/beginner/examples_autograd/two_layer_net_custom_function.html))
- `backward` is tested against dense linear algebra, results within imposed relative tolerance

### Benchmark
 [new run](https://gitlab.pasqal.com/emulation/rydberg-atoms/benchmarks_github_gitlab/-/jobs/1467498/artifacts/file/benchmarks/emu_sv_cpu_gpu/results/emu_sv_runtimes.png) vs [old run](https://gitlab.pasqal.com/emulation/rydberg-atoms/benchmarks_github_gitlab/-/jobs/1463051/artifacts/file/benchmarks/emu_sv_cpu_gpu/results/emu_sv_runtimes.png) of the `emu_sv_cpu_gpu` benchmark show same runtime and memory consumption in the logs below.

For 20 qubits - GPU
new
```
step = 448/450, RSS = 260.182 MB, Δt = 0.014 s
step = 449/450, RSS = 260.182 MB, Δt = 0.014 s
step = 450/450, RSS = 260.182 MB, Δt = 0.077 s
```
old
```
step = 448/450, RSS = 260.182 MB, Δt = 0.014 s
step = 449/450, RSS = 260.182 MB, Δt = 0.014 s
step = 450/450, RSS = 260.182 MB, Δt = 0.074 s
```

For 20 qubits - CPU
new
```
step = 448/450, RSS = 1326.564 MB, Δt = 0.189 s
step = 449/450, RSS = 1326.564 MB, Δt = 0.157 s
step = 450/450, RSS = 1326.564 MB, Δt = 0.191 s
```
old
```
step = 448/450, RSS = 1300.072 MB, Δt = 0.195 s
step = 449/450, RSS = 1300.072 MB, Δt = 0.201 s
step = 450/450, RSS = 1300.072 MB, Δt = 0.211 s
```

### Minor changes
- small fix in `test_dense_operator.py ` to get rid of a warning in tests

### Disscusion
- testing also with `gradcheck`
I propose I open a separate issue for it